monte carlo + part conclusion
i was able to do the part of the conclusion that I know, at least.  I skipped rewiring and i'll be back to it, it's just a lot.  documentation is done for the time being.  gonna finish up stuff on the code end and then move over to describing the code and its unit tests